requirements.txt
Django==3.1.7
clear-cache==1.0.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
gunicorn==20.0.4
Jinja2==2.11.3
Pillow==8.1.0
psycopg2-binary==2.8.6
